--- src/Cartas.css
@@ -1,3 +1,0 @@
-.deck {
-    display: flex;
-}
