+++ src/components/Home.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react'
+import { Link } from 'react-router-dom';
+
+export default class Home extends Component {
+  render() {
+    return (
+      <div>
+        <Link to='/game'>Play</Link>
+        <Link to='/about'>About</Link>
+      </div>
+    )
+  }
+}
