+++ src/styles/Game.css
@@ -1,0 +1,35 @@
+body {
+    background-color: #016d29;
+  }
+  
+  main {
+    display: flex;
+    flex-flow: column;
+    align-items: center;
+  }
+  
+  img {
+    width: 100px;
+  }
+  
+  .dealer-card {
+    display: flex;
+    flex-flow: column;
+    align-items: center;
+  }
+  
+  .player-card {
+    display: flex;
+    flex-flow: column;
+    align-items: center;
+  }
+  
+  .play-buttons {
+    margin: 25px;
+  }
+  
+  button {
+    width: 100px;
+    margin: 7px;
+  }
+
